backend/requirements.txt
fastapi
uvicorn
moviepy
python-multipart
requests
edge-tts
librosa
google-generativeai
gputil
psutil
yake
